Updated binary again
Only apparent bug is the off by 1 error, the python script generates 1
more line of data than this, for now I'm going to leave it until I have
more time. If we are in compact data from that usually means high sample
rate so a single data point shouldn't skew the results really.